agent: Rename httpd to web
